[core] Move integrations to their own packages
